# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CheckTimezoneRequest:
    properties:
      timezone:
        type: string
    required:
    - timezone
    type: object
  dto.CheckTimezoneResponse:
    properties:
      current:
        type: string
      detected:
        type: string
      match:
        type: boolean
    type: object
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  dto.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  dto.UpdateTimezoneRequest:
    properties:
      timezone:
        type: string
    required:
    - timezone
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Productivity tracking backend API
  title: Pulse API
  version: "1.0"
paths:
...
  /user/timezone:
    put:
      consumes:
      - application/json
      description: Updates the authenticated user's timezone
      parameters:
      - description: New timezone
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTimezoneRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update timezone
      tags:
      - User
  /user/timezone/check:
    post:
      consumes:
      - application/json
      description: Compares the detected timezone with the user's stored timezone
      parameters:
      - description: Detected timezone
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CheckTimezoneRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CheckTimezoneResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Check timezone
      tags:
      - User
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer <your_token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"